requirements.txt
streamlit
pandas
requests
python-dotenv
openpyxl
snowflake-connector-python
python-Levenshtein
pytest
